requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
psycopg2-binary
requests
pytest
pytest-django
